requirements.txt
django
psycopg2-binary

gunicorn
django-extensions
django-storages[google]
django-environ
google-cloud-secret-manager
django-phonenumber-field[phonenumbers]
